Skeleton for app_context.c and gui bindings filled out.
