+++ src/app/app.h
@@ -23,6 +23,7 @@
 #include <jnxc_headers/jnxunixsocket.h>
 #include "app_context.h"
 #include "app_auth_callbacks.h"
+#include "app_gui_bindings.h"
 
 #define CMDLEN 64
 
@@ -43,7 +44,6 @@ void app_quit_message();
 char *app_get_session_message();
 int app_code_for_command_with_param(char *command, jnx_size cmd_len,
     char **oparam);
-void app_create_gui_session(session *s, app_context_t *app_context);
 void app_list_active_peers(app_context_t *context);
 peer *app_peer_from_input(app_context_t *context,char *param);
