+++ src/app/app_context.c
@@ -1,12 +1,12 @@
 /*
  * =====================================================================================
  *
- *       Filename:  app_gui_bindings.h
+ *       Filename:  app_context.c
  *
  *    Description:  
  *
  *        Version:  1.0
- *        Created:  17/05/2015 08:32:22
+ *        Created:  17/05/2015 08:46:18
  *       Revision:  none
  *       Compiler:  gcc
  *
@@ -15,12 +15,7 @@
  *
  * =====================================================================================
  */
-
-#ifndef __APP_GUI_BINDINGS_H__
-#define __APP_GUI_BINDINGS_H__
-
+#include <stdlib.h>
 #include "app_context.h"
 
-void app_start_gui_for_session(session *s, app_context_t *app_context);
 
-#endif // __APP_GUI_BINDINGS_H__
